refact the code for more organisation
